move events funcs to util
